requirements.txt
langchain==1.2.13
langchain-community==0.4.1
langchain-openai==1.1.12
sqlalchemy==2.0.48
pymysql==1.1.2
python-dotenv==1.2.2
